+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 2019 Apr ??
 
 *   Improved error handling for underlying driver/fw/hw errors.
+*   Added command-line option to specify the joulescope device name.
 
 
 ## 0.2.7